install in Makefile, better handling of args
Using **kwargs so that I *should* just need a single function to handle all of the requests to the WCF. This will really only matter if I include more than just pulling game data by tournament from the database, but for now it's nice. I probably could use it for authentication too...

Test that.